Orientation stage of edge detection tentatively working
Updates still needed (potentially), decision to be made on that.
This update adds the needed functionality to measure the pixel spans
of a given potato, as well as implementing the ability to combine the
results of multiple processing parameters sets.

A "final" list of parameters for processing needs to be completed, and
can be stored in a local list for the script in a future update.
Additionally, looping can be optimized/eliminated in future versions
to improve speed if needed.