Add scripts for camera control & dependency to init
Two files were added, one for each camera.
There is a lot of customization that can be done,
but initial testing at least can be done with them as-is.

Each script is essentially a one-liner with comments, so they
can be converted to python using the subprocess module present
in other areas already (such as file i/o)