Batch update - Excel manipulation
This commit removes some unnecessary folders and files, it adds
new pictures of potatoes to use for testing, adds a comment
on changes need to be made (search Holden TODO through the git),
and updates the excel manipulation functions in the file_io folder.
The data_storage_functions.py script is not complete. There are some
hiccups involved with appending data (adding rows) to an existing
excel sheet. If the data was all formatted as a csv, the process
would be relatively simple. Currently, there is a speed/complexity
tradeoff decision that needs to be made.