Fix camera script bugs, remove unneeded scripts, add note
This commit does a few things related to the file_io folder.
It fixes an issue with the camera controls (from python, .sh
files were working as desired). With the python functions working
along with that fix, the associated .sh files are removed. Additionally,
a similar issue was present with the mount/unmount functions.
Finally, adds a note as to how to fully deal with the slow excel file
reading problem in the file_io folder (holden_note.txt)